Prototype persistent the GitHub payload
Without this patch we don't actually do anything with the JSON payload
sent from GitHub.  This patch addresses that problem by doing some basic
processing of the GitHub payload.

This is a very basic proof of concept.  It simply writes the JSON
payload to a plain file and reads it again.

No attempt at thread safety is made.  I assume the app runs with a
shared filesystem.  It may not.