Add example json payload for a pull request
Without this patch we don't have any example fixture data for the data
that Github sends to the webhook when a pull request is opened.  This is
a problem because it's hard to know what we're working with.

This patch adds a file containing the string that is posted to the
webhook and made available in `params['payload']` inside of the Sinatra
application handler.