Fix warning already initialized constant AuthPolicy
Without this patch applied, running the tests in watchr constantly gives
the following warning on the output:

    web.rb:58: warning: already initialized constant AuthPolicy

This patch fixes the problem by checking if the constant is already set
to the desired value.  If it is not, it deletes the constant and then
immediately re-defines it to the correct value.  This completely avoids
the warning at the cost of silently changing the internal behavior of
the Trello gem.