Update code to reflect renamed files
Without this patch the renamed file structure does not match the code.
This is a problem because the app will not deploy properly.

This patch addresses the problem by adjusting the `$LOAD_PATH`,
introducing a new spec_helper file to help load code, and updating the
Procfile which starts the application process (aka dyno in Heroku
terminology).