Queue up Pull Request jobs at /event/pull_request
Without this patch we are not creating and storing jobs that are
executable by DJ.  This patch addresses the problem by introducing new
models for the Pull Request data, and associating these models with a
PullRequestJob instance.  This instance is then written to the databases
when post data includes the payload parameter routed to
/event/pull_request.

This patch also adds two rake jobs.  `api:run` and `api:pull_request`.
These tasks start a local copy of the web server using foreman and thin.
This requires the heroku toolbelt.  The `api:pull_request` task uses the
system curl command to an HTTP POST request similar to how Github would
make one when a new pull request is created.

This patch adds a queue_job helper method.  This method is intended to
be available using `helpers PuppetLabs::SinatraDJ`.  Any object that
responds to `perform` can be queued with queue_job(obj).

In addition, this patch initializes an in-memory SQLite database for
spec testing.

This patch also switches to config.ru and thin