Make /event/pull_request respond with 202 and JSON data
Without this patch the response from the service when posting the pull
request data does not contain any information.  This is a problem
because the foreign service doesn't have any indication what happened
with the web hook.

This patch addresses the problem by returning a 202 HTTP status code.
This code means: "The request has been accepted for processing, but the
processing has not been completed. The request might or might not
eventually be acted upon, as it might be disallowed when processing
actually takes place."

In addition, serialized JSON data is returned in the body of the
response.  This data contains the job id, priority, and the time the job
has been created for processing.