Create card in Trello only if they don't already exist
Without this patch the delayed jobs don't actually create the cards on
Trello.  This patch addresses the problem by implementing some basic
logic.  If a card with a matching title already exists on the board the
job does nothing.  If the card does not exist with the same title the
card is created.