Enable workless auto-scaler for worker process
Without this patch workers are not started.  With this patch applied the
Heroku API is used to automatically scale the number of worker dynos
listed in the Procfile.

When a new job is queued, a worker process will automatically be created
by setting the scale factor for the worker process to 1.  When the
worker is finished with the job it will automatically terminate itself
by setting the scale factor to 0.

These are equivalent to:

    heroku scale worker=0

This should result in considerable cost savings since the worker
processes will not persist when the queue is empty.