Update README with migration and workless configuration
Without this patch the README isn't up to date regarding the actual
steps necessary to get the Puppet webhook service up and running.  This
is a problem because there are some undocumented steps, like migrating
the database and setting up the configuration variables in Github for
the shared secret and JSON content type.

This patch addresses the problem by reviewing all of the quick start
steps and making sure they're up to date with the current state of the
code.