Only act when a pull request is opened
Without this patch the hook will create a card that does not exist or is
archived when any pull request event is received from Github.  This is
a problem because pull requests that are closed or synchronized should
not re-create cards that have been moved.

This patch addresses the problem by having the Sinatra web application
check the JSON payload and only queue the pull request job if the event
action is "opened"