Make logging work in Heroku using STDERR
Without this patch none of the logging generated by the Sinatra
application or the Delayed Job are showing up in the output of `heroku
logs`.  This is a problem because we need to be able to see what's
happening with the application and the job processing.  The core problem
is that logs in Heroku are captured off the STDOUT and STDERR streams,
while the default logging for Active Record goes to a file on disk.

This patch addresses the problem by logging to STDOUT which should be
caught by Heroku.