Add Sinatra ActiveRecord Extension
Without this patch it's not super easy to store models in the provided
Heroku PostgreSQL database.  This is a problem because we need to store
every request and payload that comes in so we can make decisions about
what to implement next.

This patch makes the Sinatra ActiveRecord Extension available for use by
the PDC app.  This is the plugin available at:
https://github.com/janko-m/sinatra-activerecord

config/database.yml should automatically get picked up.  If it does not,
the database may be configured as per:

https://github.com/janko-m/sinatra-activerecord/wiki/Alternative-database-setup