Add PuppetLabs::PullRequestController class
Without this class most of the logic regarding what to do with the pull
request data is embedded directly in the route.  This is a problem
because the route block is already quite large and a dedicated
controller class seems like a more appropriate place to put these
conditional behaviors and the behaviors of each event.