Fix spec test failures
Without this patch the spec tests have been failing due to the
reorganization of the code.  This patch addresses the problem by
updating all of the spec tests.

The return code from 204 with JSON content as a behavior has also been
patched since Rack forces no body to exist if 204 is returned as a
status code.  Instead we return 200 with the JSON message.