Docs: Mention foreman start to run code without a credit card
Without this patch it seems like the code requires a credit card to run.
This is a problem because a credit card is a barrier to getting started
with the code.  While not nearly as a good an experience, the code may
be run locally on a machine using foreman start.  This patch updates the
README to mention this.