Limit jobs to 10 minutes and enable heroku logging
Without this patch the default limit for jobs is in effect.  This could
cause the worker process to exceed the number of inexpensive hours
available in Heroku.  This patch addresses the problem by limiting jobs
to 10 minutes.  In addition, the retry count is extended to 5 from 3.

This patch also makes sure the worker process sends log messages to
STDERR so that Heroku picks them up and makes them available via `heroku
logs`.