Add Job classes for Closed and Reopened Pull Requests
Without this patch the events generated when a pull request is closed or
reopened are ignored by the application.  This is a problem because I
need an easy way to see what happens when a job is queued up from a
logging point of view.  Closing and reopening a pull request is an easy
way to generate some events.

This patch addresses the problem by subclassing the main PullRequestJob
with specific PullRequestReopenedJob and PullRequestCLosedJob classes.
The classes override the perform instance method called by the
DelayedJob worker process.