Rename Job classes to be Trello Specific
Without this patch all of the classes dealing with performing heavy
lifting are named generically.  PullRequestJob and IssueJob and Job for
example.  These classes are fairly focused and
specific to performing the job of creating a new card on a Trello list.

This is a problem because we do not make it clear to the reader the
specific nature of these delayed job classes.

This patch addresses the problem be renaming many of the classes to
include a Trello specific name.