Fix issue handling constant exception
Without this patch the rename from IssueJob to TrelloIssueJob missed one
constant in the IssueController class.  This is a problem because issues
were not actually being handled.  The root cause of the problem is that
there was no test coverage for this behavior so I overlooked it while
performing the refactoring.

This patch addresses the problem by adding very basic test coverage for
an opened issue.  This does cause the fixed code path to be exercised.