Add Apache 2.0 License
Without this patch the license of this software is not clear.  This is a
problem because we'd very much like to see people contribute new
functionality and additional integrations.  They may be hesitant to do
so if they're unsure of the software license.  This patch addresses the
problem by making the license clear in the README.md and the LICENSE
file.

[ci skip] because no behavior is affected