Store the payload data in PullRequest#message
Without this patch the original payload data of the pull request event
is not stored along with the delayed job.  This makes it difficult to
change the behavior of the job classes in a manner that is compatible
with already stored jobs.

This patch addresses the problem by storing the entire de-serialized
object as an instance variable of the PullRequest instance.  The
instance variable data is accessible through the message method.
Methods that implement job behavior should directly read the message
data and not other instance variables in order to be backwards
compatible with the data stored in the jobs table.