Add a due date to cards created
Without this patch the cards created by this app have no due date
associated with them.  This is a problem because I'd like to make sure
all cards have actions by 2 PM the business day following when they were
created.

This patch addresses the problem by checking if
TRELLO_SET_TARGET_RESPONSE_TIME has the value "true" and if so, sets a
due date on the newly created card.