Add timezone support and update README
Without this patch it is not clear how to get the application running in
Heroku to correctly set the due date of a created card to 2 PM for the
time zone the business is actually in.  This behavior is subject to the
TZ environmental variable which is set using `heroku config:add
TZ=America/Los_Angeles`

This patch addresses the problem by adding a configuration section to
the README which mentions the timezone configuration setting.

In addition, the patch also updates the README with information about
recently added features such as searching multiple boards to prevent
card duplication and archiving cards when a pull request is closed.