Add author information to card
Without this patch, while looking at the Trello board it's difficult to
see who opened a pull request.  This is a problem because I often need
to find a card that a particular user is talking about.

This patch addresses the problem by adding an avatar image of the
contributor, their full name, their company, and their email to the
description of the card.  In addition, their full name is included in
the card title.

As part of this change the job needs to communicate with Github to
obtain the full data about the github login who sent the pull request.

Here's what the card looks like:

![Trello Card](http://goo.gl/sNMa3)

Finally, this change set also changes the behavior of how the app
matches up events to cards that may exist on the trello board.  Without
this patch applied the entire string of the expected card title is
searched for and matched.  This is a problem because the same card
should be found if the issue or pull request title changes.  This patch
addresses the problem by restricting the card identifier to be strictly
based on the repository name and issue number.  Each job class is
expected to implement a new `card_identifier` method.  The application
will search card titles and perform a substring match.