(maint) Remove commented out expected_pr_body
The examples no longer set expectations against the entire body of the
Trello card created from a pull request because this has proven fragile.
Instead, they're using substring matching to check expected behavior.
As part of this change I commented out this method and it should be
removed to clean up the code.

This patch removes the commented method.