Use the Github account name if the full name is not available
Without this patch the author_name method of the GithubMix module will
return nil if the full name of the account is not available.  This is a
problem because the full name is often not available, for example:
https://github.com/brunoleon does not have an email address, full name,
or company, and the card titles look weird, like:

    (PR hiera/123) Interpolate hash keys like already done for strings. []

This patch addresses the problem by returning the account name if the
full name is not available, so the card title looks better.  For
example:

    (PR hiera/123) Interpolate hash keys like already done for strings. [brunoleon]