(maint) Move ENV testing into the base trello job spec
Without this patch the behavior of #save_settings is tested in the
higher level PullRequestJob spec files.  This is a problem because it
looks like there is no test coverage of the BaseTrelloJob class.

This patch moves the testing of #save_settings into the newly
established file for exercising behavior on the BaseTrelloJob class.