Add BaseTrelloJob#archive_card? instance method
Without this patch jobs that speak to trello don't have an easy way to
determine if they should archive the card or not.  This is a problem
because we'd like to make this behavior easy for end users to configure.

This patch addresses the problem by implementing the archive_card?
instance method which reads the ARCHIVE_CARD environment variable via
BaseTrelloJob#save_settings when the job is initially created and
before saving it to the database.