Only archive cards if ARCHIVE_CARD=true or yes
This patch finishes the work wiring up delayed jobs to the original
user-configured environment.  With this patch applied the
TrelloPullRequestClosedJob class will only archive a card if the
application environment was configured with ARCHIVE_CARD=true or
ARCHIVE_CARD=yes at the time the job was originally created.

The values are case insensitive.