Fix unhandled exception trying to authenticate against Travis
Without this patch JSON data posted to the events endpoint will cause an
unhandled exception when trying to check for travis authentication.
This exception is caused by the request data not containing the expected
structure of a Travis event.

The method that checks if a request is an authentic request from Travis
is expected to simply return false for requests that do not originate
from travis.  The method itself should probably be cleaned up to make
this more robust, but this patch is sufficient to avoid the 500 server
error.