Save events before authenticating them
Without this patch events that cause unhandled exceptions in the
authentication logic will not be saved into the Events table for later
inspection.  This makes it difficult to see what incoming events are
causing the exception in the first place.

This patch changes the behavior of the system to store events to the
database prior to authenticating them.  There is some security concern
with this, as an attacker could overflow the events table with garbage
data, but this events table is primarily used for troubleshooting and
not for providing actual service, so the risk is worth the value of
having the troubleshooting information available.