Order the summary sections
Without this patch the order of the sections in the card summary gist is
not consistent.  This is a problem because the email should contain the
same ordering of the sections Merged, Not Merged, Other each week and
it's a pain to manually copy and paste them.

This patch addresses the problem by creating a new hash and inserting
the keys and values in key sorted order.