Remove jira stats from /event/github route
The PullRequestController is entangled with the output of the pull
request app, so changing the output in the controller breaks the sinatra
app. Since we don't actually use this information and the request
routing will need some work later, we can revert this to make the tests
pass, and address this at a later date.