Centralize pull request event lookup
All pull request event handlers need to be able to look up issues
related to pull requests in a related manner. The open event handler
could link a pull request to a referenced issue but the close and reopen
handlers would only look up issues by ID. This commit pulls all of that
behavior into a parent class so there can be a consistent lookup
approach.